feat: expose time axis across parsed waveform formats
- Added a TimeAxis abstraction (x_origin, x_increment, sample_rate) and a WfmFile::time_axis() method covering DS1000Z, DS2000, DS4000, DHO800/DHO1000, and Tektronix ISF. Origins follow the centered-trigger convention used by RigolWFM.
- Extended the ISF parser to capture XINCR and XZERO; upgraded the DHO time fields from f32 to f64 to preserve precision over deep-memory captures.
- Python API: added x_origin, x_increment, sample_rate properties and a get_time_axis() helper that returns a float64 NumPy array. Returns None for formats whose time fields are not yet parsed (DS1000E, Tektronix WFM).
- CLI: the info subcommand now prints sample rate, capture duration, and time origin in human-readable SI units. The convert subcommand writes a leading time column in CSV by default and a structured NumPy dtype in NPY (np.load(...)['time'], np.load(...)['CH1']); pass --no-time to opt out.
- Fixed model_number and firmware_version string parsing to terminate at the first NUL byte instead of carrying trailing non-printable bytes. Surfaced when reading DS2000 and DS4000 captures whose 20-byte string fields are not zero-padded after the NUL terminator.

+++ crates/wfm_oxide/src/dho.rs
@@ -8,8 +8,8 @@ const FILE_HEADER_SIZE: usize = 24;
 const BLOCK_HEADER_SIZE: usize = 12;
 const ADC_MIDPOINT: f32 = 32768.0;
 
-const DHO1000_TICK_S: f32 = 1e-8;
-const DHO800_TICK_S: f32 = 8e-10;
+const DHO1000_TICK_S: f64 = 1e-8;
+const DHO800_TICK_S: f64 = 8e-10;
 
 const BLOCK_TYPE_DHO800_PARAMS: u16 = 5;
 const BLOCK_TYPE_SETTINGS: u16 = 6;
@@ -30,8 +30,8 @@ pub struct DhoHeader {
     pub n_pts_per_ch: usize,
     pub n_ch: usize,
     pub data_start: usize,
-    pub x_increment: f32,
-    pub x_origin: f32,
+    pub x_increment: f64,
+    pub x_origin: f64,
 }
 
 impl DhoHeader {
@@ -226,7 +226,7 @@ fn find_data_section(
     data: &[u8],
     blocks_end: usize,
     is_dho800: bool,
-) -> Result<(usize, usize, usize, f32, f32)> {
+) -> Result<(usize, usize, usize, f64, f64)> {
     let mut offset = blocks_end;
     while offset < data.len() && data[offset] == 0 {
         offset += 1;
@@ -257,8 +257,8 @@ fn find_data_section(
         x_increment_raw = 1;
     }
     let tick_s = if is_dho800 { DHO800_TICK_S } else { DHO1000_TICK_S };
-    let x_increment = (x_increment_raw as f64 * tick_s as f64) as f32;
-    let x_origin = -(n_pts_per_ch as f32 / 2.0) * x_increment;
+    let x_increment = x_increment_raw as f64 * tick_s;
+    let x_origin = -(n_pts_per_ch as f64 / 2.0) * x_increment;
 
     Ok((n_pts_per_ch, n_ch, offset + 40, x_increment, x_origin))
 }
